Replace client side stats page with server side one
