+++ src/app/stats/components/DelegationHistoryGraph.tsx
