+++ src/app/stats/utils/fetchMint.ts
